revised to deal with clashes in the format
Having the use of _ in both the emphasis and also the format of the tag names caused intractable collisions in interpretation.  Backed off and dropped _ in the emphasis of the tag, but left it in place in the tag name.  Bc that's more how programmers think, they want _ in their names sometimes to declare certain spaces, and we don't need both forms of emphasis.  There's still a collision there bc GFM will interpret the _ as emphasis whereas Riccy will not.  Some more thought required.